requirements.txt
absl-py==1.4.0
addict==2.4.0
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
asynctest==0.13.0
attrs==23.1.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
cmake==3.26.3
cnstd==1.2.2
codecov==2.1.13
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.0.7
coverage==7.2.5
cycler==0.11.0
ensemble-boxes==1.0.9
exceptiongroup==1.1.1
filelock==3.12.0
flake8==6.0.0
flatbuffers==23.3.3
fonttools==4.39.3
frozenlist==1.3.3
fsspec==2023.4.0
google-auth==2.17.3
google-auth-oauthlib==1.0.0
grpcio==1.54.0
humanfriendly==10.0
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
iniconfig==2.0.0
interrogate==1.5.0
isort==4.3.21
Jinja2==3.1.2
kiwisolver==1.4.4
kwarray==0.6.12
lightning-utilities==0.8.0
lit==16.0.2
llvmlite==0.40.0
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
mccabe==0.7.0
mdurl==0.1.2
mmcv-full==1.2.7
mmengine==0.7.3
mmpycocotools==12.0.3
model-index==0.1.11
mpmath==1.3.0
multidict==6.0.4
networkx==3.1
numba==0.57.0
numpy==1.24.3
oauthlib==3.2.2
onnx==1.13.1
onnxruntime==1.14.1
opencv-python==4.7.0.72
openmim==0.3.7
ordered-set==4.1.0
packaging==23.1
pandas==2.0.1
pdf2image==1.16.3
Pillow==9.5.0
pluggy==1.0.0
Polygon3==3.0.9.1
protobuf==3.20.3
py==1.11.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyclipper==1.3.0.post4
pycocotools==2.0.6
pycodestyle==2.10.0
pyflakes==3.0.1
Pygments==2.15.1
PyMuPDF==1.22.2
pyparsing==3.0.9
pytesseract==0.3.10
pytest==7.3.1
python-dateutil==2.8.2
pytorch-lightning==2.0.2
pytz==2023.3
PyYAML==6.0
requests==2.29.0
requests-oauthlib==1.3.1
rich==13.3.5
rsa==4.9
scipy==1.10.1
seaborn==0.12.2
shapely==2.0.1
six==1.16.0
sympy==1.11.1
tabulate==0.9.0
tensorboard==2.12.3
tensorboard-data-server==0.7.0
termcolor==2.3.0
terminaltables==3.1.10
toml==0.10.2
tomli==2.0.1
torch==2.0.0
torchaudio==0.7.2
torchmetrics==0.11.4
torchvision==0.15.1
tqdm==4.65.0
triton==2.0.0
typing_extensions==4.5.0
tzdata==2023.3
ubelt==1.2.4
Unidecode==1.3.6
urllib3==1.26.15
Werkzeug==2.3.3
xdoctest==1.1.1
yapf==0.33.0
yarl==1.9.2
zipp==3.15.0
